@@ -5,12 +5,12 @@
   "scripts": {
     "start": "BABEL_DISABLE_CACHE=1 NODE_ENV=development nodemon index.js",
     "start:prod": "NODE_ENV=production node index.js",
-    "lint": "eslint client server",
-    "lint-fix": "eslint client server --fix",
+    "lint": "eslint client server --ext .jsx --ext .js",
+    "lint-fix": "eslint client server --ext .jsx --ext .js --fix",
     "build": "NODE_ENV=production webpack --config webpack.config.prod.js --progress",
     "test": "echo \"Error: no test specified\" && exit 1",
     "fetch-examples": "node fetch-examples.js",
-    "postinstall" : "git submodule update --remote --recursive"
+    "postinstall": "git submodule update --remote --recursive"
   },
   "main": "index.js",
   "author": "Cassie Tarakajian",
@@ -20,27 +20,25 @@
     "url": "git+https://github.com/catarak/p5.js-web-editor.git"
   },
   "devDependencies": {
-    "babel-eslint": "^6.1.0",
+    "babel-eslint": "^7.1.1",
     "babel-loader": "^6.2.4",
     "babel-plugin-transform-react-constant-elements": "^6.8.0",
     "babel-plugin-transform-react-inline-elements": "^6.8.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.2.6",
-    "babel-polyfill": "^6.8.0",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-es2015-native-modules": "^6.9.2",
     "babel-preset-react": "^6.5.0",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-react-optimize": "^1.0.1",
     "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.8.0",
     "chunk-manifest-webpack-plugin": "^0.1.0",
     "css-loader": "^0.23.1",
     "cssnano": "^3.7.1",
-    "eslint": "^2.13.1",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-import": "^1.9.2",
-    "eslint-plugin-jsx-a11y": "^1.5.3",
-    "eslint-plugin-react": "^5.2.2",
+    "eslint": "^3.14.0",
+    "eslint-config-airbnb": "^14.0.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^3.0.2",
+    "eslint-plugin-react": "^6.9.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
     "node-sass": "^3.7.0",
@@ -49,14 +47,8 @@
     "postcss-focus": "^1.0.0",
     "postcss-loader": "^0.9.1",
     "postcss-reporter": "^1.3.3",
-    "redux-devtools": "^3.3.1",
-    "redux-devtools-dock-monitor": "^1.1.1",
-    "redux-devtools-log-monitor": "^1.0.11",
     "sass-loader": "^3.2.0",
     "style-loader": "^0.13.1",
-    "webpack": "^1.14.0",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-hot-middleware": "^2.10.0",
     "webpack-manifest-plugin": "^1.1.0"
   },
   "engines": {
@@ -67,6 +59,8 @@
     "async": "^2.0.0",
     "axios": "^0.12.0",
     "babel-core": "^6.8.0",
+    "babel-polyfill": "^6.8.0",
+    "babel-register": "^6.8.0",
     "bcrypt-nodejs": "0.0.3",
     "blob-util": "^1.2.1",
     "body-parser": "^1.15.1",
@@ -86,7 +80,7 @@
     "htmlhint": "^0.9.13",
     "js-beautify": "^1.6.4",
     "jsdom": "^9.8.3",
-    "jshint": "^2.9.2",
+    "jshint": "^2.9.4",
     "lodash": "^4.16.4",
     "loop-protect": "git+https://git@github.com/catarak/loop-protect.git",
     "moment": "^2.14.1",
@@ -106,6 +100,9 @@
     "react-router": "^2.6.0",
     "react-split-pane": "^0.1.44",
     "redux": "^3.5.2",
+    "redux-devtools": "^3.3.1",
+    "redux-devtools-dock-monitor": "^1.1.1",
+    "redux-devtools-log-monitor": "^1.0.11",
     "redux-form": "^5.3.3",
     "redux-thunk": "^2.1.0",
     "request": "^2.76.0",
@@ -113,6 +110,9 @@
     "s3-policy": "^0.2.0",
     "shortid": "^2.2.6",
     "srcdoc-polyfill": "^0.2.0",
+    "webpack": "^1.14.0",
+    "webpack-dev-middleware": "^1.6.1",
+    "webpack-hot-middleware": "^2.10.0",
     "xhr": "^2.2.1"
   }
 }
