@@ -1,38 +1,21 @@
-import * as ActionTypes from '../../../constants';
 import { browserHistory } from 'react-router';
 import axios from 'axios';
+import objectID from 'bson-objectid';
+import * as ActionTypes from '../../../constants';
 import { showToast, setToastText } from './toast';
 import { setUnsavedChanges,
   justOpenedProject,
   resetJustOpenedProject,
-  setProjectSavedTime,
-  resetProjectSavedTime,
   showErrorModal } from './ide';
-import moment from 'moment';
 
 const ROOT_URL = location.href.indexOf('localhost') > 0 ? 'http://localhost:8000/api' : '/api';
 
-export function getProject(id) {
-  return (dispatch, getState) => {
-    const state = getState();
-    dispatch(justOpenedProject());
-    if (state.ide.justOpenedProject) {
-      dispatch(resetProjectSavedTime());
-    }
-    axios.get(`${ROOT_URL}/projects/${id}`, { withCredentials: true })
-      .then(response => {
-        dispatch({
-          type: ActionTypes.SET_PROJECT,
-          project: response.data,
-          files: response.data.files,
-          owner: response.data.user
-        });
-        dispatch(setUnsavedChanges(false));
-      })
-      .catch(response => dispatch({
-        type: ActionTypes.ERROR,
-        error: response.data
-      }));
+export function setProject(project) {
+  return {
+    type: ActionTypes.SET_PROJECT,
+    project,
+    files: project.files,
+    owner: project.user
   };
 }
 
@@ -40,6 +23,21 @@ export function setProjectName(name) {
   return {
     type: ActionTypes.SET_PROJECT_NAME,
     name
+  };
+}
+
+export function getProject(id) {
+  return (dispatch, getState) => {
+    dispatch(justOpenedProject());
+    axios.get(`${ROOT_URL}/projects/${id}`, { withCredentials: true })
+      .then((response) => {
+        dispatch(setProject(response.data));
+        dispatch(setUnsavedChanges(false));
+      })
+      .catch(response => dispatch({
+        type: ActionTypes.ERROR,
+        error: response.data
+      }));
   };
 }
 
@@ -53,9 +51,10 @@ export function saveProject(autosave = false) {
     formParams.files = [...state.files];
     if (state.project.id) {
       axios.put(`${ROOT_URL}/projects/${state.project.id}`, formParams, { withCredentials: true })
-        .then(() => {
+        .then((response) => {
           dispatch(setUnsavedChanges(false));
-          dispatch(setProjectSavedTime(moment().format()));
+          console.log(response.data);
+          dispatch(setProject(response.data));
           dispatch({
             type: ActionTypes.PROJECT_SAVE_SUCCESS
           });
@@ -85,9 +84,9 @@ export function saveProject(autosave = false) {
         });
     } else {
       axios.post(`${ROOT_URL}/projects`, formParams, { withCredentials: true })
-        .then(response => {
+        .then((response) => {
           dispatch(setUnsavedChanges(false));
-          dispatch(setProjectSavedTime(moment().format()));
+          dispatch(setProject(response.data));
           browserHistory.push(`/${response.data.user.username}/sketches/${response.data.id}`);
           dispatch({
             type: ActionTypes.NEW_PROJECT,
@@ -107,7 +106,7 @@ export function saveProject(autosave = false) {
             }
           }
         })
-        .catch(response => {
+        .catch((response) => {
           if (response.status === 403) {
             dispatch(showErrorModal('staleSession'));
           } else {
@@ -130,7 +129,7 @@ export function autosaveProject() {
 export function createProject() {
   return (dispatch) => {
     axios.post(`${ROOT_URL}/projects`, {}, { withCredentials: true })
-      .then(response => {
+      .then((response) => {
         browserHistory.push(`/${response.data.user.username}/sketches/${response.data.id}`);
         dispatch({
           type: ActionTypes.NEW_PROJECT,
@@ -165,14 +164,38 @@ export function newProject() {
   return resetProject();
 }
 
+function generateNewIdsForChildren(file, files) {
+  const newChildren = [];
+  file.children.forEach((childId) => {
+    const child = files.find(childFile => childFile.id === childId);
+    const newId = objectID().toHexString();
+    child.id = newId;
+    child._id = newId;
+    newChildren.push(newId);
+    generateNewIdsForChildren(child, files);
+  });
+  file.children = newChildren; // eslint-disable-line
+}
+
 export function cloneProject() {
   return (dispatch, getState) => {
     dispatch(setUnsavedChanges(false));
     const state = getState();
-    const formParams = Object.assign({}, { name: `${state.project.name} copy` }, { files: state.files });
+    const newFiles = state.files.map((file) => { // eslint-disable-line
+      return { ...file };
+    });
+
+    const rootFile = newFiles.find(file => file.name === 'root');
+    const newRootFileId = objectID().toHexString();
+    rootFile.id = newRootFileId;
+    rootFile._id = newRootFileId;
+    generateNewIdsForChildren(rootFile, newFiles);
+    // const newFiles = state.files;
+    const formParams = Object.assign({}, { name: `${state.project.name} copy` }, { files: newFiles });
     axios.post(`${ROOT_URL}/projects`, formParams, { withCredentials: true })
-      .then(response => {
+      .then((response) => {
         browserHistory.push(`/${response.data.user.username}/sketches/${response.data.id}`);
+        console.log(response.data);
         dispatch({
           type: ActionTypes.NEW_PROJECT,
           project: response.data,
@@ -198,4 +221,3 @@ export function hideEditProjectName() {
     type: ActionTypes.HIDE_EDIT_PROJECT_NAME
   };
 }
-
