@@ -1,13 +1,14 @@
-import Project from '../models/project';
-import User from '../models/user';
 import archiver from 'archiver';
 import request from 'request';
 import moment from 'moment';
+import Project from '../models/project';
+import User from '../models/user';
 
 
 export function createProject(req, res) {
   if (!req.user) {
-    return res.status(403).send({ success: false, message: 'Session does not match owner of project.' });
+    res.status(403).send({ success: false, message: 'Session does not match owner of project.' });
+    return;
   }
 
   let projectValues = {
@@ -17,50 +18,63 @@ export function createProject(req, res) {
   projectValues = Object.assign(projectValues, req.body);
 
   Project.create(projectValues, (err, newProject) => {
-    if (err) { return res.json({ success: false }); }
+    if (err) {
+      res.json({ success: false });
+      return;
+    }
     Project.populate(newProject,
       { path: 'user', select: 'username' },
       (innerErr, newProjectWithUser) => {
-        if (innerErr) { return res.json({ success: false }); }
-        return res.json(newProjectWithUser);
+        if (innerErr) {
+          res.json({ success: false });
+          return;
+        }
+        res.json(newProjectWithUser);
       });
   });
 }
 
 export function updateProject(req, res) {
-  Project.findById(req.params.project_id, (err, project) => {
+  Project.findById(req.params.project_id, (findProjectErr, project) => {
     if (!req.user || !project.user.equals(req.user._id)) {
-      return res.status(403).send({ success: false, message: 'Session does not match owner of project.' });
+      res.status(403).send({ success: false, message: 'Session does not match owner of project.' });
+      return;
     }
-    // if (req.body.updatedAt && moment(req.body.updatedAt) < moment(project.updatedAt)) {
-    //   return res.status(409).send({ success: false, message: 'Attempted to save stale version of project.' });
-    // }
+    if (req.body.updatedAt && moment(req.body.updatedAt) < moment(project.updatedAt)) {
+      res.status(409).send({ success: false, message: 'Attempted to save stale version of project.' });
+      return;
+    }
     Project.findByIdAndUpdate(req.params.project_id,
       {
         $set: req.body
+      },
+      {
+        new: true
       })
       .populate('user', 'username')
-      .exec((err, updatedProject) => {
-        if (err) {
-          console.log(err);
-          return res.json({ success: false });
+      .exec((updateProjectErr, updatedProject) => {
+        if (updateProjectErr) {
+          console.log(updateProjectErr);
+          res.json({ success: false });
+          return;
         }
         if (updatedProject.files.length !== req.body.files.length) {
           const oldFileIds = updatedProject.files.map(file => file.id);
           const newFileIds = req.body.files.map(file => file.id);
           const staleIds = oldFileIds.filter(id => newFileIds.indexOf(id) === -1);
-          staleIds.forEach(staleId => {
+          staleIds.forEach((staleId) => {
             updatedProject.files.id(staleId).remove();
           });
-          updatedProject.save((innerErr) => {
+          updatedProject.save((innerErr, savedProject) => {
             if (innerErr) {
               console.log(innerErr);
-              return res.json({ success: false });
+              res.json({ success: false });
+              return;
             }
-            return res.json(updatedProject);
+            res.json(savedProject);
           });
         }
-        return res.json(updatedProject);
+        res.json(updatedProject);
       });
   });
 }
@@ -77,15 +91,17 @@ export function getProject(req, res) {
 }
 
 export function deleteProject(req, res) {
-  Project.findById(req.params.project_id, (err, project) => {
+  Project.findById(req.params.project_id, (findProjectErr, project) => {
     if (!req.user || !project.user.equals(req.user._id)) {
-      return res.status(403).json({ success: false, message: 'Session does not match owner of project.' });
+      res.status(403).json({ success: false, message: 'Session does not match owner of project.' });
+      return;
     }
-    Project.remove({ _id: req.params.project_id }, (err) => {
-      if (err) {
-        return res.status(404).send({ message: 'Project with that id does not exist' });
+    Project.remove({ _id: req.params.project_id }, (removeProjectError) => {
+      if (removeProjectError) {
+        res.status(404).send({ message: 'Project with that id does not exist' });
+        return;
       }
-      return res.json({ success: true });
+      res.json({ success: true });
     });
   });
 }
@@ -100,7 +116,7 @@ export function getProjects(req, res) {
       });
   } else {
     // could just move this to client side
-    return res.json([]);
+    res.json([]);
   }
 }
 
@@ -108,18 +124,18 @@ export function getProjectsForUser(req, res) {
   if (req.params.username) {
     User.findOne({ username: req.params.username }, (err, user) => {
       if (!user) {
-        return res.status(404).json({ message: 'User with that username does not exist.' });
+        res.status(404).json({ message: 'User with that username does not exist.' });
+        return;
       }
       Project.find({ user: user._id }) // eslint-disable-line no-underscore-dangle
         .sort('-createdAt')
         .select('name files id createdAt updatedAt')
-        .exec((err, projects) => res.json(projects));
+        .exec((innerErr, projects) => res.json(projects));
     });
   } else {
     // could just move this to client side
-    return res.json([]);
+    res.json([]);
   }
-  return null;
 }
 
 function buildZip(project, req, res) {
@@ -127,7 +143,6 @@ function buildZip(project, req, res) {
   const rootFile = project.files.find(file => file.name === 'root');
   const numFiles = project.files.filter(file => file.fileType !== 'folder').length;
   const files = project.files;
-  const projectName = project.name;
   let numCompletedFiles = 0;
 
   zip.on('error', (err) => {
@@ -140,27 +155,25 @@ function buildZip(project, req, res) {
   function addFileToZip(file, path) {
     if (file.fileType === 'folder') {
       const newPath = file.name === 'root' ? path : `${path}${file.name}/`;
-      file.children.forEach(fileId => {
+      file.children.forEach((fileId) => {
         const childFile = files.find(f => f.id === fileId);
         (() => {
           addFileToZip(childFile, newPath);
         })();
       });
-    } else {
-      if (file.url) {
-        request({ method: 'GET', url: file.url, encoding: null }, (err, response, body) => {
-          zip.append(body, { name: `${path}${file.name}` });
-          numCompletedFiles += 1;
-          if (numCompletedFiles === numFiles) {
-            zip.finalize();
-          }
-        });
-      } else {
-        zip.append(file.content, { name: `${path}${file.name}` });
+    } else if (file.url) {
+      request({ method: 'GET', url: file.url, encoding: null }, (err, response, body) => {
+        zip.append(body, { name: `${path}${file.name}` });
         numCompletedFiles += 1;
         if (numCompletedFiles === numFiles) {
           zip.finalize();
         }
+      });
+    } else {
+      zip.append(file.content, { name: `${path}${file.name}` });
+      numCompletedFiles += 1;
+      if (numCompletedFiles === numFiles) {
+        zip.finalize();
       }
     }
   }
