@@ -1,6 +1,8 @@
 import mongoose from 'mongoose';
-const Schema = mongoose.Schema;
+
 const bcrypt = require('bcrypt-nodejs');
+
+const Schema = mongoose.Schema;
 
 const userSchema = new Schema({
   name: { type: String, default: '' },
@@ -40,7 +42,7 @@ userSchema.pre('save', function checkPassword(next) { // eslint-disable-line con
   });
 });
 
-userSchema.virtual('id').get(function () {
+userSchema.virtual('id').get(function idToString() {
   return this._id.toHexString();
 });
 
